requirements.txt
wrapt
natsort
websocket-client